Almost fucked everything up repushing the entire project to fix it. added a countercomponent to replace the live and score components. added a howtoplay screen.

+++ Minigin/SubjectComponent.cpp
@@ -30,7 +30,7 @@ void engine::SubjectComponent::RemoveObserver(const std::shared_ptr<Observer>&)
 
 void engine::SubjectComponent::Notify(std::weak_ptr<GameObject> entity, Event event)
 {
-	for (int i{ 0 }; i < m_pObservers.size(); ++i)
+	for (size_t i{ 0 }; i < m_pObservers.size(); ++i)
 	{
 		m_pObservers[i]->OnNotify(entity, event);
 	}
